+++ CHANGELOG.rst
@@ -7,6 +7,9 @@ v4.3.0
 
 - An MPRIS-based trigger is not available. This will set caffeine to inhibit
   based on the state of a currently running media player.
+- The pulseaudio integration is now disabled by default. It is deprecated and
+  will be removed in the 5.0.0 release. The MPRIS integration yields more
+  consistent results and should be used instead.
 - ``caffeine kill`` and `caffeine start --kill`` been removed. You can Quit
   caffeine via the tray icon menu. These commands are redundant, there's plenty
   of ways to find and kill a process (e.g.: ``ps aux | grep caffeine`` and